+++ packages/plugins/plugin-table/src/hooks/index.ts
@@ -4,4 +4,3 @@
 
 export * from './useTableModel';
 export * from './useTableMenuController';
-export * from './useTableIntialisation';
